# Performance enhancement: avoid repeated slow loads of the SIAF